modified tags, posts and threads, fixed hiding/deleting

+++ backend/routes/users.js
@@ -18,11 +18,14 @@ function getTime(){
 }
 router.get('/find/:loginUser',isAdmin, async (req, res) => {
     console.log(`INFO User ${req.user.login} is trying to find user ${req.params.loginUser} ${getTime()}`)
-    const login = req.params.loginUser
     try{
-        const foundUser = await User.findOne({login:login}).skip()
-        console.log(`INFO User ${req.user.login} successfully found user ${req.params.loginUser} ${getTime()}`)
-        return res.status(200).json({user:foundUser})
+        const login = req.params.loginUser
+        const user = await User.findById(req.user.id)
+        if(user && user.isAdmin){
+            const foundUser = await User.findOne({login:login}).skip()
+            console.log(`INFO User ${req.user.login} successfully found user ${req.params.loginUser} ${getTime()}`)
+            return res.status(200).json({user:foundUser})
+        }
     }
     catch (err){
         console.log(`ERROR ${err} ${getTime()}`)
@@ -30,13 +33,16 @@ router.get('/find/:loginUser',isAdmin, async (req, res) => {
     }
 });
 router.get(`/:page/:limit`,isAdmin, async (req, res) => {
-    const page = req.params.page
-    const limit = req.params.limit
     console.log(`INFO User ${req.user.login} is requesting the list of all users ${getTime()}`)
     try{
-        const users =  await User.find().skip((page-1)*limit).limit(Number(limit)).sort({createdAt:1})
-        console.log(`INFO User ${req.user.login} was granted the list of all users ${getTime()}`)
-        return res.status(200).json({users:users})
+        const page = req.params.page
+        const limit = req.params.limit
+        const user = await User.findById(req.user.id)
+        if(user.isAdmin){
+            const users =  await User.find().skip((page-1)*limit).limit(Number(limit)).sort({createdAt:1})
+            console.log(`INFO User ${req.user.login} was granted the list of all users ${getTime()}`)
+            return res.status(200).json({users:users})
+        }
     }
     catch(err){
         console.log(`ERROR ${err} ${getTime()}`)
@@ -46,9 +52,10 @@ router.get(`/:page/:limit`,isAdmin, async (req, res) => {
 router.delete('/:idUser',isAdmin, async (req, res) => {
     console.log(`INFO User ${req.user.login} is trying to delete user ${req.params.idUser} ${getTime()}`)
     try{
+        const userRequest = await User.findById(req.user.id)
         const user = await User.findById(req.params.idUser)
-        if(!user.isAdmin){
-            await User.findByIdAndDelete(req.params.idUser)
+        if(!user.isAdmin && userRequest.isAdmin){
+            await User.findByIdAndDelete(user._id)
             console.log(`INFO User ${req.params.idUser} was succesfully deleted ${getTime()}`)
             return res.json({message:"User deleted",status:200})
         }
@@ -62,13 +69,14 @@ router.delete('/:idUser',isAdmin, async (req, res) => {
 
 router.post('/patch/:idUser', async (req, res) => {
     console.log(`INFO User ${req.user.login} is trying to patch user ${req.params.idUser} ${getTime()}`)
-    const id = req.params.idUser;
     try{
-        if(req.user.id === req.params.idUser){
+        const id = req.params.idUser;
+        const user = await User.findById(req.user.id)
+        if(user._id.equals(req.params.idUser)){
             if(req.body.login){
                 const users = await User.find({login:req.body.login})
                 if(users.length===0){
-                    await User.findByIdAndUpdate(req.params.idUser,
+                    await User.findByIdAndUpdate(user._id,
                     { $set: {login:req.body.login}},
                     { new:true, runValidators:true})
                     console.log(`INFO User's ${req.body.login} LOGIN was patched successfully ${getTime()}`)
@@ -78,11 +86,11 @@ router.post('/patch/:idUser', async (req, res) => {
                     return res.status(500).json({message:"Username already taken"})
                 }
             }
-            else if(req.body.password && req.body.repeatedPassword){
+            else if(!req.body.password.empty() && !req.body.repeatedPassword.empty() && req.body.password===req.body.repeatedPassword){
                 let salt = crypto.randomBytes(Number(SALT_BITS));
-                const users = await User.find({login:req.body.login})
+                const user = await User.find({login:req.body.login})
                     crypto.pbkdf2(req.body.password, salt, 310000, 32, HASH_FUNCTION, async (err, hashedPassword) => {
-                        const user = await User.findByIdAndUpdate(req.params.idUser,
+                        const user = await User.findByIdAndUpdate(user._id,
                             { $set: {password:hashedPassword}},
                             { new:true, runValidators:true}
                         )
@@ -104,8 +112,11 @@ router.post('/patch/:idUser', async (req, res) => {
 router.get(`/toBeAccepted`,isAdmin, async (req, res,next) => {
     console.log(`INFO Admin ${req.user.login} is requesting users to be accepted ${getTime()}`)
     try{
-        const users = await User.find({isAcceptedByAdmin:false,isAdmin:false})
-        return res.json({users:users,status:200})
+        const user = await User.findById(req.user.id)
+        if(user.isAdmin){
+            const users = await User.find({isAcceptedByAdmin:false,isAdmin:false})
+            return res.json({users:users,status:200})
+        }
     }
     catch(err){
         console.log(`ERROR ${err} ${getTime()}`)
@@ -115,12 +126,15 @@ router.get(`/toBeAccepted`,isAdmin, async (req, res,next) => {
 router.post(`/toBeAccepted/:id`,isAdmin, async (req, res,next) => {
     console.log(`INFO Admin ${req.user.login} is accepting user ${req.params.id} ${getTime()}`)
     try{
-        const user = await User.findByIdAndUpdate(req.params.id,
+        const userRequest = await User.findById(req.user.id)
+        if(userRequest.isAdmin){
+            const user = await User.findByIdAndUpdate(req.params.id,
             { $set: {isAcceptedByAdmin:true}},
             { new:true, runValidators:true})
-        req.app.get('io').to('admins').emit('userAccepted',user)
-        req.app.get('io').to('adminsChat').emit('adminMessage',`INFO - User ${user.login} was accepted!`)
-        return res.json({user:user,status:200})
+            req.app.get('io').to('admins').emit('userAccepted',user)
+            req.app.get('io').to('adminsChat').emit('adminMessage',`INFO - User ${user.login} was accepted!`)
+            return res.json({user:user,status:200})
+        }
     }
     catch(err){
         console.log(`ERROR ${err} ${getTime()}`)
@@ -130,10 +144,13 @@ router.post(`/toBeAccepted/:id`,isAdmin, async (req, res,next) => {
 router.post(`/notToBeAccepted/:id`,isAdmin, async (req, res,next) => {
     console.log(`INFO Admin ${req.user.login} is not accepting user ${req.params.id} ${getTime()}`)
     try{
-        const user = await User.findByIdAndDelete(req.params.id)
-        req.app.get('io').to('admins').emit('userNotAccepted',user)
-        req.app.get('io').to('adminsChat').emit('adminMessage',`INFO - User ${user.login} was denied!`)
-        return res.json({user:user,status:200})
+        const userRequest = await User.findById(req.user.id)
+        if(userRequest.isAdmin){
+            const user = await User.findByIdAndDelete(req.params.id)
+            req.app.get('io').to('admins').emit('userNotAccepted',user)
+            req.app.get('io').to('adminsChat').emit('adminMessage',`INFO - User ${user.login} was denied!`)
+            return res.json({user:user,status:200})
+        }
     }
     catch(err){
         console.log(`ERROR ${err} ${getTime()}`)
@@ -143,11 +160,14 @@ router.post(`/notToBeAccepted/:id`,isAdmin, async (req, res,next) => {
 router.post(`/giveAdmin/:id`,isAdmin,async(req,res,next)=>{
     console.log(`INFO Admin ${req.user.login} is trying to give admin to user ${req.params.id} ${getTime()}`)
     try{
-        const user = await User.findByIdAndUpdate(req.params.id,
+        const userRequest = await User.findById(req.user.id)
+        if(userRequest.isAdmin){
+            const user = await User.findByIdAndUpdate(req.params.id,
             { $set: {isAdmin:true}},
             { new:true, runValidators:true})
-        req.app.get('io').to('admins').emit('adminAdded',user)
-        return res.json({user:user,status:200})
+            req.app.get('io').to('admins').emit('adminAdded',user)
+            return res.json({user:user,status:200})
+        }
     }
     catch(err){
         console.log(`ERROR ${err} ${getTime()}`)
